feat: refactor DevtronLicenseCard component and introduce types for license management
