+++ src/Shared/Components/DevtronLicenseCard/index.tsx
@@ -1,1 +1,3 @@
 export { default as DevtronLicenseCard } from './DevtronLicenseCard'
+export { parseDevtronLicenseDTOIntoLicenseCardData } from './utils'
+export * from './types'
