+++ src/Shared/Helpers.tsx
@@ -51,12 +51,9 @@ import {
 import {
     AggregationKeys,
     BorderConfigType,
-    DevtronLicenseCardProps,
-    DevtronLicenseDTO,
     GitTriggers,
     IntersectionChangeHandler,
     IntersectionOptions,
-    LicenseStatus,
     Nodes,
     PreventOutsideFocusProps,
     TargetPlatformItemDTO,
@@ -1069,38 +1066,6 @@ export const deriveBorderRadiusAndBorderClassFromConfig = ({
 
 export const getClassNameForStickyHeaderWithShadow = (isStuck: boolean, topClassName = 'dc__top-0') =>
     `dc__position-sticky ${topClassName} dc__transition--box-shadow ${isStuck ? 'dc__box-shadow--header' : ''}`
-
-const getDevtronLicenseStatus = ({
-    ttl,
-    reminderThreshold,
-}: Pick<DevtronLicenseDTO, 'ttl' | 'reminderThreshold'>): LicenseStatus => {
-    if (ttl < 0) {
-        return LicenseStatus.EXPIRED
-    }
-
-    if (ttl < reminderThreshold * 24 * 60 * 60) {
-        return LicenseStatus.REMINDER_THRESHOLD_REACHED
-    }
-
-    return LicenseStatus.ACTIVE
-}
-
-export const parseDevtronLicenseDTOIntoLicenseCardData = <isCentralDashboard extends boolean = false>(
-    licenseDTO: DevtronLicenseDTO<isCentralDashboard>,
-    currentUserEmail?: isCentralDashboard extends true ? string : never,
-): DevtronLicenseCardProps => {
-    const { isTrial, expiry, ttl, reminderThreshold, organisationMetadata, license, claimedByUserDetails } =
-        licenseDTO || {}
-
-    return {
-        enterpriseName: organisationMetadata?.name || '',
-        expiryDate: expiry ? moment(expiry).format(DATE_TIME_FORMATS['DD/MM/YYYY']) : '',
-        ttl,
-        licenseStatus: getDevtronLicenseStatus({ ttl, reminderThreshold }),
-        isTrial,
-        ...(currentUserEmail === claimedByUserDetails?.email ? { licenseKey: license } : { licenseSuffix: license }),
-    }
-}
 
 export const clearCookieOnLogout = () => {
     document.cookie = `${TOKEN_COOKIE_NAME}=; expires=Thu, 01-Jan-1970 00:00:01 GMT;path=/`
