+++ src/Shared/types.ts
@@ -1070,26 +1070,3 @@ export type DevtronLicenseDTO<isCentralDashboard extends boolean = false> = Devt
           })
 
 export type CountryISO2Type = ParsedCountry['iso2']
-
-export enum LicenseStatus {
-    ACTIVE = 'ACTIVE',
-    EXPIRED = 'EXPIRED',
-    REMINDER_THRESHOLD_REACHED = 'REMINDER_THRESHOLD_REACHED',
-}
-
-export type DevtronLicenseCardProps = {
-    enterpriseName: string
-    expiryDate: string
-    ttl: number
-    licenseStatus: LicenseStatus
-    isTrial: boolean
-} & (
-    | {
-          licenseKey: string
-          licenseSuffix?: never
-      }
-    | {
-          licenseKey?: never
-          licenseSuffix: string
-      }
-)
